- Addressed another batch of warnings produced by the MSVC compiler and static analysis tools. In addition to more of the same kind of warnings as in the previous commit, this also includes corrections for signed/unsigned comparisons, non-virtual destructors, non-explicit constructors, missing copy constructors, float comparisons, empty switches and const-correctness. Due to the nature of the changes involved (e.g. changed method signatures), there is a slight chance that there are some unwanted side-effects. On the bright side though, the following issues have been identified and fixed as a direct result of the warning clean-up:  > DEFMSG_NPC_PET_WAGE_COST not showing the hire cost.  > CItem::Use_SpyGlass not searching the entire area for land and just checking one title.  > CWorldThread::m_FreeOffset being initialised with -1 which would result in either an access violation when used to lookup against m_FreeUIDs, or even disable the use of m_FreeUIDs altogether and only use the slower free uid lookup.  > Unicode text not being adjusted properly by SuppressCapitals setting, since COUNTOF() was being used on a pointer and returning the wrong value.  > Possible memory leak due to PacketHouseDesign's destructor not necessarily being invoked.  > Missing copy constructor and assignment operator on CResourceDef lead to the reference count for some resources becoming inaccurate when they are ever used since the compiler-generated constructor+operator would not update the reference count.
- Expanded _NEWGUILDSYSTEM checks so that methods used in the 'old' guild system are completely removed rather than just their contents being cut out.
- Related to the CResourceDef issue, added copy constructors and assignment operators to almost any class which would accept them. These are marked as private and without a body so that the compiler will throw an error if used, to prevent accidental copying and assignment. If these are ever needed for real then a body will need to be created.
- Removed AbstractSphereThread::allocateString(void) as it would not have worked correctly due to the temporary TemporaryString object's destructor marking the allocated string as 'unused' (the result being something similar to Str_GetTemp).
- Removed the 'using' directive for the std namespace since it is causing problems between our definition of 'abs' and 'std::abs' when building with GCC. Everything is already properly qualified anyway so the statement is not needed.

+++ 0.56.b/common/CQueue.cpp
@@ -24,7 +24,7 @@ const BYTE * CQueueBytes::RemoveDataLock() const
 	return m_Mem.GetData();
 }
 
-void CQueueBytes::RemoveDataAmount( int iSize )
+void CQueueBytes::RemoveDataAmount( size_t iSize )
 {
 	ADDTOCALLSTACK("CQueueBytes::RemoveDataAmount");
 	// use up this read data. (from the start)
@@ -34,12 +34,12 @@ void CQueueBytes::RemoveDataAmount( int iSize )
 	memmove( m_Mem.GetData(), m_Mem.GetData()+iSize, m_iDataQty );
 }
 
-BYTE * CQueueBytes::AddNewDataLock( int iLen )
+BYTE * CQueueBytes::AddNewDataLock( size_t iLen )
 {
 	ADDTOCALLSTACK("CQueueBytes::AddNewDataLock");
 	// lock the queue to place this data in it.
 	ASSERT(iLen>=0);
-	int iLenNew = m_iDataQty + iLen;
+	size_t iLenNew = m_iDataQty + iLen;
 	if ( iLenNew > m_Mem.GetDataLength() )
 	{
 		// re-alloc a bigger buffer. as needed.
@@ -51,13 +51,13 @@ BYTE * CQueueBytes::AddNewDataLock( int iLen )
 	return( m_Mem.GetData() + m_iDataQty );
 }
 
-void CQueueBytes::AddNewDataFinish( int iLen )
+void CQueueBytes::AddNewDataFinish( size_t iLen )
 {
 	// The data is now ready.
 	m_iDataQty += iLen;
 }
 
-void CQueueBytes::AddNewData( const BYTE * pData, int iLen )
+void CQueueBytes::AddNewData( const BYTE * pData, size_t iLen )
 {
 	ADDTOCALLSTACK("CQueueBytes::AddNewData");
 	// Add new data to the end of the queue.
